+++ database/migrations/vitess/0000000099-create-email_list_subscribers-table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `email_list_subscribers` (
   `id` bigint PRIMARY KEY auto_increment,
-  `email_list_id` integer not null REFERENCES `email_lists`(`id`),
-  `subscriber_id` integer not null REFERENCES `subscribers`(`id`),
+  `email_list_id` bigint not null REFERENCES `email_lists`(`id`),
+  `subscriber_id` bigint not null REFERENCES `subscribers`(`id`),
   `status` ENUM('subscribed', 'unsubscribed', 'pending', 'bounced') not null default 'subscribed',
   `source` varchar(100) default 'api',
   `subscribed_at` datetime,
